Split up select type into subpages for each renderType
The select type can have 5 different renderTypes. For each renderType a
number of properties exist and are listed.

This used to be all on one page which made it difficult to determine
which properties exist for one rendertype and when scrolling it could
easily happen that you were looking at a property for a different renderType.

This patch makes the following changes:

- Put properties and examples for each renderType on a subpage
- Add an table of contents of all properties at the beginning of the
  properties section
- Change the ordering in the examples to list the code directly under
  the screenshot and not all screenshots and then all codeblocks (as already
  done with the checkbox page).

Resolves: #258
Related: #253